Delete this test, it's quite broken.
This tests to see if your installed RT has the relevant callbacks, but
it can't possibly work on modern RTs (in fact I doubt it worked properly
on 3.8 since it never looked in plugins for the RTIR files).  It also
hardcoded a bunch of /opt/rt3 paths despite having support for knowing
RT's real installed path in other tests.

It also doesn't understand that autohandler callbacks moved to lib/ in
later 3.8s.

The utility in fixing this greatly outweighs the utility in having the
test.