Switch to modern RT::Test that can load/init RTIR for us
Change to using as many RT::IR::Test::Web methods as possible
Update messages where needed

Parts of this are still really inconsistent with changes in RTIR, not
sure where the blame lies (in this module or in the RTIR changes).
Still fails tests about halfway through, but this is as far as I got.